Fixed build problem with prerendering of cms glossaries page
